requirements.txt
fastapi==0.103.2
uvicorn[standard]==0.22.0
google-cloud-bigquery==3.11.4
pydantic==1.10.13